Create curl_post_check_status_null_response.sh
This bash script runs a cURL command, particulary a graphql, as a POST request. The purpose is to check the status code whether that graphql for a particular "id" will give successful response or not. 

No need to capture the response.

And write the status code for each id into an output file